tests/wpt/tests/html/canvas/element/canvas-host/2d.canvas.host.initial.reset.gradient.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.canvas.host.initial.reset.gradient</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.canvas.host.initial.reset.gradient</h1>
<p class="desc">Resetting the canvas state does not invalidate any existing gradients</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Resetting the canvas state does not invalidate any existing gradients");
_addTest(function(canvas, ctx) {

  canvas.width = 50;
  var g = ctx.createLinearGradient(0, 0, 100, 0);
  g.addColorStop(0, '#0f0');
  g.addColorStop(1, '#0f0');
  canvas.width = 100;
  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  ctx.fillStyle = g;
  ctx.fillRect(0, 0, 100, 50);
  _assertPixel(canvas, 50,25, 0,255,0,255);

});
</script>
